Driver_obj {DriverFuse}	R Documentation

Driver obj

Description

It reports the structural variants and copy number variation profile of input fusion that is responsible for their formation.

Usage

Driver_obj("MYH9-EIF3D")

Arguments

x
fusion gene

y
svc.df used for mapping fusion genes coordinates to structural variants

z
CNV segementation table used for mapping fusion gene coordinates to there segmentation profile

Value

‘Driver_obj’ creates object of input fusion gene based on their mapping structural variant and copy number variation.

Examples

Driver_obj("MYH9-EIF3D")

[Package DriverFuse version 0.1.0 Index]