Driver_plot {DriverFuse}	R Documentation

Driver_plot

Description

It plots the copy number variation profile and structural variants that are mapping to genomic coordinates of input fusion genes.

Usage

Driver_plot("MYH9-EIF3D")

Arguments

x
fusion gene

Value

Plots fusion gene coordinates with their respective matching structural variant and copy number profile of user dataset.

Examples

Driver_result("MYH9-EIF3D")

[Package DriverFuse version 0.1.0 Index]