input {DriverFuse}	R Documentation

Preparing input files

Description

This function creates localization of breakpoints with respect to known genes. The annotation identifies small segmental variants overlapping with genes.

Usage

results_svc <- svc.break.annot(svc, svc.seg.size = 200000, genome.v="hg19",upstr = 100000, verbose=FALSE)

Arguments

x
svc.df used for mapping fusion genes coordinates to structural variants

Value

This function incorporates structural variant with gene annotation.

[Package DriverFuse version 0.1.0 Index]