input_cnv {DriverFuse}	R Documentation

Preparing input files

Description

It creates input copy number variation file for a user sample data that is used for mapping fusion gene to find out out whether it is “Driver” or “Passenger” fusion gene.

Usage

input_cnv(x)

Arguments

x
CNV segementation table used for mapping fusion gene coordinates to there segmentation profile

Value

Preparing input copy number variation file of our sample data form mapping input fusion gene.

Examples

x = read.csv("path")
cnv = input_cnv(x)

[Package DriverFuse version 0.1.0 Index]