input_svc {DriverFuse}	R Documentation

Preparing input files

Description

It creates input structural varaint file for a user sample data that is used for mapping fusion gene to find out whether it is “Driver” or “Passenger” fusion gene.

Usage

input_svc(x)

Arguments

x
svc.df used for mapping fusion genes coordinates to structural variants

Value

Preparing structural variant file of our sample data form mapping input fusion gene.

Examples

x = read.csv("path")
svc = input_svc(x)

[Package DriverFuse version 0.1.0 Index]